Fix searching IP by name, incorrect unregistration MYSQLQuery, register/unregister PgQuery. Add using internal DNS cache, source documentation updated.

+++ trunk/src/odbcquery.cpp
@@ -286,7 +286,7 @@ bool ODBCQuery::fetch ()
   return ok;
 }
 
-int ODBCQuery::affectedRows ()
+long ODBCQuery::affectedRows ()
 {
   long err;
   SQLINTEGER count;
